1st runnable version of model with deeplsd attractionfield
