+++ gluefactory/models/extractors/jpldd/__init__.py
@@ -1,2 +1,2 @@
 # jpldd = Joint Point Line Detector (And) Descriptor
-from .joint_line_point_extractor import JointPointLineDetectorDescriptor
+from .joint_point_line_extractor import JointPointLineDetectorDescriptor
